+++ thrift_tools/tests/util.py
@@ -8,3 +8,7 @@ _resources_dir = os.path.join(
 
 def get_pcap_path(capture_file):
     return os.path.join(_resources_dir, '%s.pcap' % (capture_file))
+
+
+def get_log_path(log_file):
+    return os.path.join(_resources_dir, '%s.logfile' % (log_file))