+++ thrift_tools/thrift_file.py
@@ -8,7 +8,10 @@ try:
 except ImportError:
     HAS_MMAP = False
 
+from thrift.transport import TTransport
+
 from .thrift_message import ThriftMessage
+from .thrift_struct import ThriftStruct
 
 
 class ThriftFile(object):
@@ -39,6 +42,9 @@ class ThriftFile(object):
             self._data = fh.read()
             self._view = memoryview(self._data)
 
+    def _data_slice(self, idx):
+        return self._view[idx:].tobytes() if self._view else self._data[idx:]
+
 
 class ThriftMessageFile(ThriftFile):
     """
@@ -52,8 +58,8 @@ class ThriftMessageFile(ThriftFile):
     """
 
     def __init__(self, file_name='-', finagle_thrift=False,
-                 read_values=False, padding=0):
-        super(ThriftMessageFile, self).__init__(file_name, read_values, True)
+                 read_values=False, padding=0, debug=False):
+        super(ThriftMessageFile, self).__init__(file_name, read_values, debug)
         self._padding = padding
         self._finagle_thrift = finagle_thrift
 
@@ -63,11 +69,8 @@ class ThriftMessageFile(ThriftFile):
             msg, hole = self._read_next(idx, len(self._data))
             if hole:
                 idx += hole[1]  # number of bytes skipped
-            idx += msg.length + self._padding
+            idx += len(msg) + self._padding
             yield msg, hole
-
-    def _data_slice(self, idx):
-        return self._view[idx:].tobytes() if self._view else self._data[idx:]
 
     def _read_next(self, start, end):
         for idx in range(start, end):
@@ -76,7 +79,56 @@ class ThriftMessageFile(ThriftFile):
                                             finagle_thrift=self._finagle_thrift,
                                             read_values=self._read_values)
                 skipped = idx - start
-                return msg, None if skipped == 0 else (msg, (start, skipped))
+                return (msg, None) if skipped == 0 else (msg, (start, skipped))
+            except Exception, ex:
+                if self._debug:
+                    print('Bad message: %s (idx=%d)' % (ex, idx))
+
+
+class ThriftStructFile(ThriftFile):
+    """
+    A file containing thrift structs. Allows iteration via standard iterator
+    protocol
+
+    Ex:
+    >> thrift_struct_file = ThriftStructFile(file_name)
+    >> for tstruct in thrift_msg_file:
+        print tstruct.as_dict()
+    """
+
+    MAX_FIELDS = 10000
+    MAX_LIST_SIZE = 10000
+    MAX_MAP_SIZE = 10000
+    MAX_SET_SIZE = 10000
+
+    def __init__(self, protocol, file_name='-', read_values=False, padding=0, debug=False):
+        super(ThriftStructFile, self).__init__(file_name, read_values, debug)
+        self._padding = padding
+        self._protocol = protocol
+
+    def __iter__(self):
+        idx = self._padding
+        while idx < len(self._data):
+            tstruct, hole = self._read_next(idx, len(self._data))
+            if hole:
+                idx += hole[1]  # number of bytes skipped
+            idx += tstruct.bytes_length + self._padding
+            yield tstruct, hole
+
+    def _read_next(self, start, end):
+        for idx in range(start, end):
+            try:
+                trans = TTransport.TMemoryBuffer(self._data_slice(idx))
+                proto = self._protocol(trans)
+                tstruct = ThriftStruct.read(
+                    proto,
+                    max_fields=ThriftStructFile.MAX_FIELDS,
+                    max_list_size=ThriftStructFile.MAX_LIST_SIZE,
+                    max_map_size=ThriftStructFile.MAX_MAP_SIZE,
+                    max_set_size=ThriftStructFile.MAX_SET_SIZE,
+                    read_values=self._read_values)
+                skipped = idx - start
+                return (tstruct, None) if skipped == 0 else (tstruct, (start, skipped))
             except Exception, ex:
                 if self._debug:
                     print('Bad message: %s (idx=%d)' % (ex, idx))